Implement access to X11 primary selection for GLFW
New functions glfwSetSelectionString() and glfwGetSelectionString()
are added similar to glfwSetClipboardString() and
glfwGetClipboardString(). These functions do nothing on platforms with
no X11-like primary selection.

This was done for mupdf-gl, which includes its own modified version of
GLFW. It might be useful in general for GLFW, if it is not deemed too
specific to the X11 platform.

Note: Only tested on Linux due to my lack of available other platforms.